feat: option for user to choose the register
