+++ README.md
@@ -1,8 +1,8 @@
 # Markdown Title Fetch for Neovim
 
-Automatically fetches the title of a webpage and formats a markdown link. The plugin creates a command `MarkdownLinkPaste` that takes the contents of the `*` register, and if they it is a url, fetches the title of the url and formats a markdown link adding it to the current buffer at the current location. If the link does not contain a webpage with a `<title>` tag, then the domain of the url will be used instead.
+Automatically fetches the title of a webpage and formats a markdown link. The plugin creates a command `MarkdownLinkPaste` that takes the contents of the register, and if they it is a url, fetches the title of the url and formats a markdown link adding it to the current buffer at the current location. If the link does not contain a webpage with a `<title>` tag, then the domain of the url will be used instead.
 
-If the contents of the '*' register is not a link, then the contents will be pasted as is.
+If the contents of the register is not a link, then the contents will be pasted as is.
 
 ## Example
 
@@ -30,7 +30,9 @@ For more information see: [GitHub - folke/lazy.nvim: 💤 A modern plugin manage
 ```lua
 {
   "https://github.com/samerickson/markdown-title-fetch.nvim",
-  opts = {},
+  opts = {
+    register = "*", -- The register that the plugin will read link from
+  },
   cmd = "MarkdownLinkPaste",
   keys = {
     { "<c-v>", "<cmd>MarkdownLinkPaste<cr>" },
